+++ options.go
@@ -34,24 +34,31 @@ type SubscribeOptions struct {
 	ExpireAt int64
 	// ChannelInfo defines custom channel information, zero value means no channel information.
 	ChannelInfo []byte
-	// Presence turns on participating in channel presence - i.e. client subscription
-	// will be visible in a current channel presence.
-	Presence bool
-	// JoinLeave enables sending Join and Leave messages for this client in channel.
-	JoinLeave bool
-	// When position is on client will additionally sync its position inside
-	// a stream to prevent message loss. The loss can happen due to at most once
-	// guarantees of PUB/SUB model. Make sure you are enabling Position in channels
-	// that maintain Publication history stream. When Position is on Centrifuge will
+	// EmitPresence turns on participating in channel presence - i.e. client
+	// subscription will emit presence updates to PresenceManager and will be visible
+	// in a channel presence result.
+	EmitPresence bool
+	// EmitJoinLeave turns on emitting Join and Leave events from the subscribing client.
+	// See also PushJoinLeave if you want current client to receive join/leave messages.
+	EmitJoinLeave bool
+	// PushJoinLeave turns on receiving channel Join and Leave events by the client.
+	// Subscriptions which emit join/leave events should have EmitJoinLeave on.
+	PushJoinLeave bool
+	// When position is on client will additionally sync its position inside a stream
+	// to prevent publication loss. The loss can happen due to at most once guarantees
+	// of PUB/SUB model. Make sure you are enabling EnablePositioning in channels that
+	// maintain Publication history stream. When EnablePositioning is on Centrifuge will
 	// include StreamPosition information to subscribe response - for a client to be
 	// able to manually track its position inside a stream.
-	Position bool
-	// Recover turns on automatic recovery for a channel. In this case client will try to
-	// recover missed messages upon resubscribe to a channel after reconnect to a server.
-	// This option also enables client position tracking inside a stream (i.e. enabling
-	// Recover will automatically enable Position option) to prevent occasional message loss.
-	// Make sure you are using Recover in channels that maintain Publication history stream.
-	Recover bool
+	EnablePositioning bool
+	// EnableRecovery turns on automatic recovery for a channel. In this case
+	// client will try to recover missed messages upon resubscribe to a channel
+	// after reconnect to a server. This option also enables client position
+	// tracking inside a stream (i.e. enabling EnableRecovery will automatically
+	// enable EnablePositioning option) to prevent occasional publication loss.
+	// Make sure you are using EnableRecovery in channels that maintain Publication
+	// history stream.
+	EnableRecovery bool
 	// Data to send to a client with Subscribe Push.
 	Data []byte
 	// RecoverSince will try to subscribe a client and recover from a certain StreamPosition.
@@ -79,31 +86,38 @@ func WithChannelInfo(chanInfo []byte) SubscribeOption {
 	}
 }
 
-// WithPresence ...
-func WithPresence(enabled bool) SubscribeOption {
+// WithEmitPresence ...
+func WithEmitPresence(enabled bool) SubscribeOption {
 	return func(opts *SubscribeOptions) {
-		opts.Presence = enabled
+		opts.EmitPresence = enabled
 	}
 }
 
-// WithJoinLeave ...
-func WithJoinLeave(enabled bool) SubscribeOption {
+// WithEmitJoinLeave ...
+func WithEmitJoinLeave(enabled bool) SubscribeOption {
 	return func(opts *SubscribeOptions) {
-		opts.JoinLeave = enabled
+		opts.EmitJoinLeave = enabled
 	}
 }
 
-// WithPosition ...
-func WithPosition(enabled bool) SubscribeOption {
+// WithPushJoinLeave ...
+func WithPushJoinLeave(enabled bool) SubscribeOption {
 	return func(opts *SubscribeOptions) {
-		opts.Position = enabled
+		opts.PushJoinLeave = enabled
 	}
 }
 
-// WithRecover ...
-func WithRecover(enabled bool) SubscribeOption {
+// WithPositioning ...
+func WithPositioning(enabled bool) SubscribeOption {
 	return func(opts *SubscribeOptions) {
-		opts.Recover = enabled
+		opts.EnablePositioning = enabled
+	}
+}
+
+// WithRecovery ...
+func WithRecovery(enabled bool) SubscribeOption {
+	return func(opts *SubscribeOptions) {
+		opts.EnableRecovery = enabled
 	}
 }